+++ src/api/factories/article-payload.api.factory.ts
@@ -1,11 +1,6 @@
+import { ArticlePayload } from '@_src/api/models/article-payload.api.model';
 import { prepareRandomArticle } from '@_src/ui/factories/article.factory';
 
-export interface ArticlePayload {
-  title: string;
-  body: string;
-  date: string;
-  image: string;
-}
 export function prepareArticlePayload(): ArticlePayload {
   const randomArticleData = prepareRandomArticle();
   return {
