+++ src/api/models/article-payload.api.model.ts
@@ -1,0 +1,6 @@
+export interface ArticlePayload {
+  title: string;
+  body: string;
+  date: string;
+  image: string;
+}
