+++ src/api/models/comment-payload.api.model.ts
@@ -1,0 +1,5 @@
+export interface CommentPayload {
+  article_id: number;
+  body: string;
+  date: string;
+}
